Started to work on simulation code.
This project will use the the GitHub repositories
listed below as references for the simulation code:
    - zschneider/falling-sand
    - Kledal/html5-sand-game

A class, called `FleckMaterial`, has been added to
represent the physical and visual properties of
particles in the simulated world. In addition, a
class to manage and update particles and other
information for the simulated world has been added.
It is called `FleckWorld`.